Add RouterSet event indexing for all registries
