+++ subgraph/abis/PortalRegistry.json
@@ -146,6 +146,19 @@
     "type": "event"
   },
   {
+    "anonymous": false,
+    "inputs": [
+      {
+        "indexed": true,
+        "internalType": "address",
+        "name": "router",
+        "type": "address"
+      }
+    ],
+    "name": "RouterSet",
+    "type": "event"
+  },
+  {
     "inputs": [
       {
         "internalType": "address[]",
