--- subgraph/config/linea-testnet.json
@@ -1,6 +1,0 @@
-{
-  "schemaRegistry": "0x0123456789abcdef0123456789abcdef01234567",
-  "moduleRegistry": "0x0123456789abcdef0123456789abcdef01234568",
-  "attestationRegistry": "0x0123456789abcdef0123456789abcdef01234569",
-  "attestationReader": "0x0123456789abcdef0123456789abcdef01234570"
-}
